File cleanup for ui/screens/ui and moving user model
