+++ app/src/main/java/com/example/secondnature/MainActivity.kt
@@ -17,7 +17,7 @@ import com.example.secondnature.ui.screens.home.HomeScreen
 import com.example.secondnature.ui.screens.post.PostScreen
 import com.example.secondnature.ui.screens.profile.ProfileScreen
 import com.example.secondnature.ui.screens.search.SearchScreen
-import ui.screens.ui.ui.theme.SecondNatureTheme
+import com.example.secondnature.ui.theme.SecondNatureTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
