+++ app/src/main/java/com/example/secondnature/viewmodel/CreateAccountViewModel.kt
@@ -7,7 +7,7 @@ import com.google.firebase.auth.FirebaseAuthUserCollisionException
 import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.tasks.await
 import android.util.Log
-import com.example.secondnature.viewmodel.User
+import com.example.secondnature.data.model.User
 
 class CreateAccountViewModel : ViewModel() {
     private val firebaseAuth: FirebaseAuth = FirebaseAuth.getInstance()
