--- app/src/main/java/ui/screens/ui/HomeScreen.kt
@@ -1,60 +1,0 @@
-package ui.screens.ui
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.ExitToApp
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.navigation.NavController
-
-data class BottomNavItem(
-    val title: String,
-    val icon: ImageVector
-)
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HomePage(navController: NavController) {
-    val navItems = listOf(
-        BottomNavItem("Home", Icons.Default.Home),
-        BottomNavItem("Profile", Icons.Default.Person),
-        BottomNavItem("Settings", Icons.Default.Settings),
-        BottomNavItem("Notifications", Icons.Default.Notifications),
-        BottomNavItem("Logout", Icons.Default.ExitToApp)
-    )
-    var selectedIndex by remember { mutableStateOf(0) }
-
-    Scaffold(
-        bottomBar = {
-            NavigationBar {
-                navItems.forEachIndexed { index, item ->
-                    NavigationBarItem(
-                        selected = index == selectedIndex,
-                        onClick = { selectedIndex = index },
-                        icon = { Icon(imageVector = item.icon, contentDescription = item.title) },
-                        label = { Text(item.title) }
-                    )
-                }
-            }
-        }
-    ) { paddingValues ->
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(paddingValues),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "Welcome to the Home Page!",
-                style = MaterialTheme.typography.headlineMedium
-            )
-        }
-    }
-}
