--- app/src/main/java/ui/screens/ui/ui/theme/Theme.kt
@@ -1,57 +1,0 @@
-package ui.screens.ui.ui.theme
-
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.sp
-
-// Define light and dark color schemes
-private val LightColorPalette = lightColorScheme(
-    primary = Color(0xFF6200EE),
-    secondary = Color(0xFF03DAC5),
-    background = Color(0xFFF1F1F1),
-    surface = Color(0xFFFFFFFF),
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black
-)
-
-private val DarkColorPalette = darkColorScheme(
-    primary = Color(0xFFBB86FC),
-    secondary = Color(0xFF03DAC5),
-    background = Color(0xFF121212),
-    surface = Color(0xFF121212),
-    onPrimary = Color.Black,
-    onSecondary = Color.Black,
-    onBackground = Color.White,
-    onSurface = Color.White
-)
-
-// Define a simple Typography
-val typography = Typography(
-    bodyLarge = TextStyle(
-        fontSize = 16.sp
-    ),
-    bodyMedium = TextStyle(
-        fontSize = 14.sp
-    ),
-    bodySmall = TextStyle(
-        fontSize = 12.sp
-    )
-)
-
-@Composable
-fun SecondNatureTheme(
-    darkTheme: Boolean = false,
-    content: @Composable () -> Unit
-) {
-    val colors = if (darkTheme) DarkColorPalette else LightColorPalette
-
-    MaterialTheme(
-        colorScheme = colors,
-        typography = typography,
-        content = content
-    )
-}
